Produce separate coappearance graphs for each chapter as well as an overall graph for the entire novel.
Code in "3k.py" will now produce a "3k-coappear.json" file for the full novel's coappearance graph, and a separate "3k-coappear-<chapter>.json" file for each chapter it can find data for in "3k.json". For now, chapters 1 and 2 are represented. The script "force.js" has been modified to be a function that takes a file path as its only argument, so the path to the data is no longer hard-coded.