Cosmetic changes to force.css and force.js.
Changed text to have white stroke/outline around it for visibility against dark backgrounds. Changed force.js so text always appears above circles in graph. Changed force.js so circle and text size are both scaled relative to the node with the largest number of links. That is, the largest circle in one graph will always be the same size as the largest circle in another graph, even if the absolute number of links to each respective node are different between the graphs. This is to avoid ginormous circles in the full-novel graph and invisibly small ones in per-chapter graphs.