Find when mousing over an item, find corresponding datum by unique id, not i.
Since the index in d3 of any given character (or other datum) in the visualization may change, use a unique id instead to search for the correct datum when mousing over a visual representation of it. Also pass the graph visualization as a parameter to appear_stack() so mousing over a character in the stack can affect the character's node in the graph.