coappear: implement dynamic resizing.
The coappearance graph and each element within the graph are now sized based on the window size, minus the space taken up by the character list sidebar on the left. Made some small adjustments to the d3 force layout parameters and node/text sizing to work better with this change.