Fix CSS bug in coappearance chapter view.
With some window sizes, the Bootstrap responsive layout would shrink width and leave the coappearance graph (which is constant-size) extending out of its well. At small sizes, the visualization is too crowded and the text too small to be readable anyway, so I just made it invisible on screens/windows below 768 px in size. I altered bootstrap-responsive.min.css to remove some of the logic for window sizes between 768 and 980 px in width (to enable the content to scroll horizontally if needed at window sizes between 768 and 1200 instead of being resized).